updated 7b

+++ 3rd Semester/Unix Shell Programming Laboratory/P7/7b.c
@@ -16,12 +16,21 @@ int main(){
 	if(dp){
 		while((dir=readdir(dp))!=NULL){
 			fd=open(dir->d_name,O_RDWR,0777);
+			if (fd == -1) {
+				perror("Error opening file");
+				continue;
+			}
 			n=lseek(fd,0,SEEK_END);
+			close(fd);
 
 			if(!n){
+				printf("Deleting empty file: %s\n", dir->d_name);
 				unlink(dir->d_name);
 			}
 		}
+		closedir(dp);
+	} else {
+		perror("Error opening directory");
 	}
 }
